mf: Introduce IMFMediaShutdownNotify for notification of media source shutdown.
Native Windows robustly handles unexpected shutdown of a media source
used in a session, and always sends MESessionStarted with an error
status, but if the native source is contained within a wrapper source
object, MESessionStarted is never sent even if BeginGetEvent() fails with
MF_E_SHUTDOWN, which should be easily handled. This implies that Windows
uses a private interface for shutdown notification, and its proper
functioning is disrupted when the source is wrapped.