quartz/dsoundrender: Use send_sample_data() to fill the buffer with silence at EOS.
Inspired by a patch by Rémi Bernon.

This allows us to clear the buffer a bit earlier than waiting for an underrun
does, and avoids the spurious FIXME message in that case. It also unifies code
paths a bit.

This change also correctly clears the buffer for 8-bit formats.