win32u: Always update the surface regions in apply_window_pos.
We need to call update_surface_region whenever the window rect changes
as it uses the visible rect to clip the regions to the visible area
of the window.

We could delegate this to the host windowing system in most cases, but
it was explicitly clipped to the visible rect in the Android driver.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=57551