winex11: Use the win32 client rect in needs_client_window_clipping.
After the change to move the sync_gl_drawable call outside the win_data
mutex in 54d82ed4, the X11 client rect used for determining whether we need
to composite/clip the drawable can be out of sync with the win32 side.
This results in offscreening when unnecessary and vice-versa.