winex11.drv: Go though WithdrawnState when transitioning from IconicState to NormalState.
When transitioning a window from IconicState to NormalState and the window is managed, go through
WithdrawnState. This is needed because Mutter doesn't unmap windows when making windows iconic/minimized
as Mutter needs to support live preview for minimized windows. So on Mutter, a window can be both
iconic and mapped. If the window is mapped, then XMapWindow() will have no effect according to the
XMapWindow() documentation. Thus we have to transition to WithdrawnState first, then to NormalState.

Fix Mass Effect Legendary Edition (1328670) starts minimized.