win32u: Don't use the current mode in the registry if it's a detached mode.
Don't use the current mode in the registry if it's a detached mode when adding virtual modes.
Otherwise, the requested/previous current mode in the registry could be a detached mode and it will
cause functions like MonitorFromRect() to return NULL, even after monitor gets turned on.

Fix Witcher 3 (292030) black screen after monitor is turned off and then on.