server: Track unbound input signaled state based on its original console.
Do not delegate to the current process's queue, either, but use the queue for
this object.

I/O acts on the current process's console, but the handle signaled state is the
same across all handles to this object, and reflects the signaled state of the
console of the process that created the object.