mf: Specify the multithreaded queue for presentation clock callbacks.
Callback invocation for timed work occurs in the standard queue. For the
session engine, this results in samples being delivered from the same
thread which manages the session and executes transforms, which in some
cases causes lag and dropped audio. In native Windows, samples may be
delivered from at least two different threads. The multithreaded work
queue responds to load, and new threads are created only when necessary.