cryptui: Copy localized name of selected store to textbox.
Match the behavior on Windows: When the CryptUIWizImport function is
called, the "Certificate store" textbox initially says "Determined by
the program" instead of having the specific name of the store passed to
the function. When the user clicks Browse to select a store, the stores
are listed by localized name. After clicking OK, the selected store's
localized name is copied into the textbox.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=58057