quartz/tests: Increase the timeout of a pause test in test_media_event.
This test failed on Linux 2 out of the last 67 times on test.winehq.org
because IMediaControl_GetState returned VFW_S_STATE_INTERMEDIATE,
evidently meaning that the command to pause had been received but not
yet carried out. Give it two seconds to complete instead of one to
ensure that it finishes.