quartz: Clamp MediaSeeking_GetCurrentPosition to the stop position.
test_graph_seeking sets the stop time to 6 seconds and then checks that
the media has stopped at that exact point in time. That test failed 10
out of the last 67 times on test.winehq.org because
IMediaSeeking_GetCurrentPosition returned a time that was 1 millisecond
after the stop time. To make the test pass consistently, clamp the
current time to the stop time.