notepad: Restructure menus and make hot-keys unique.
Some of the hot-keys, such as “Edit→Cut” and “Edit→Time/Date” had
identical “T” hot-keys, as well as “Search→Search” and “Search→Search
Next” both being on “S”.

With this change, the menus are restructured to match the Windows XP
Notepad menus, in item names, capitalization, and hot-keys.  Notably
the usability issue with conflicting hot-keys is eliminated.  Users
familiar with the Microsoft Windows Notepad menus and their hot-keys
should be able to use Wine Notepad without changing any workflow.