windows.media.speech/tests: Make refcount tests results deterministic.
GetResults on IInspectable async operation is supposed to clear the
async reference too but we don't do that yet. Releasing the async first
makes the refcount non-deterministic as the task may hold onto the
results for a bit longer, this reverts this part from the commit below
and flags the refcount test as todo_wine.

Fixes: 82771d3343bffca73873de4029149de0197d88c9