dbghelp: Don't dupe some exports on 64bit compilation.
For a 32bit DLL, dbghelp exports both the 32bit and the 64bit
variant of some APIs (eg. StackWalk, SymLoadModule...).

For a 64bit DLL, only the 64bit variant is implemented and
exported (the two names point to the same address in the export
table).

This patch:
- uses the same function for both names (as native does) in .spec file
  for 64bit architecture
- removes the 32bit variant from 64bit compilation,
- adapts also the corresponding import:s in imagehlp from dbghelp.

This mostly fixes 64bit apps, getting eg "StackWalk" address with
GetProcAddress() and expecting a 64bit code path.
(reported & tested by Stefan).

Signed-off-by: Eric Pouech <epouech@codeweavers.com>