ntdll: Mark IsBadStringPtr functions as noinline.
Without this, the compiler may inline them into dispatch_exception. When compiler
exceptions are enabled, this causes a handler to be set up for dispatch_exception,
which is unintended. A side effect of this is that EXCEPTION_NESTED_CALL may be
cleared when a nested exception is dispatched.

This fixes test_nested_exception test on aarch64 when compiler exceptions are used.