ntdll: Work around AMD SYSRET SS descriptor behavior on FreeBSD.
On AMD CPUs the cached SS descriptor attributes could be invalid after
return to the userspace with SYSRET, which doesn't matter in the
long mode, but would easily crash programs in the compatibility mode.
Reloading %ss avoids that.

Note that Linux already has X86_BUG_SYSRET_SS_ATTRS as the workaround
for this issue.