winex11: Serialize managed window config change requests.
When using a window manager, config requests are going through the WM
and their ConfigureNotify might be received with a different serial from
the request that triggered them.

We still want to override past and transient configs with our desired
state as much as possible, and we ignore the received values if some
changes were delayed, and request them instead.