widl: Do not write "const" modifiers for _PARAM_STRUCT fields.
The issue comes from per-field structure copy that we generate
on 32-bit. For non-pointer fields like "const int" this breaks
compilation when assigning to a const field. The best solution
would be to remove local struct variable completely and pass
first argument address as a stack address. This is what midl
does. My understanding is that doing it midl way used to break
on some compilers, and we have to keep the copy. Once we encounter
pointer type that needs such copy, we'll have to patch it more
to cast the const away for example.

Such type modification matches what midl does for structures in _Thunk's,
so the fix itself is correct.

Signed-off-by: Nikolay Sivov <nsivov@codeweavers.com>