d3dcompiler/tests: Avoid precision issues in the trigonometry tests.
Previously, depending on how the sin() was calculated, we could get a
small number or exactly 0, which unfortunately can be pretty far
between in terms of ULPs. We don't really want to test edge-case
accuracy with trig functions in the first place, so just nudge the
angle away from exact values.