d3dx9_36/tests: Use correct row pitch for D3DXLoadSurfaceFromMemory.
The source rect is set to be 2x2, and sizeof(pixdata) is used as the row pitch. Which means the
second row is located beyond the end of pixdata.

Tests added back in be8afaff686e, commit message doesn't seem to indicate this was trying to test
out-of-bound read. So use the correct row pitch here.