amstream/tests: Avoid calling GetAllocatorRequirements() from DecideAllocator().
Apparently amstream's reconnection will attempt to autoplug if the types don't
match. We don't implement this behaviour yet.

This isn't particularly relevant for the tests, because none of the autoplugged
filters successfully connect.

It causes a test failure here, however, because we attempt to call
IMemInputPin::GetAllocatorRequirements() on various autoplugged filters, and not
all of them return E_NOTIMPL.

Oddly, the failure is intermittent. Apparently the relevant filter is not always
autoplugged, for unclear reasons.