kernel32/tests: SetFileInfo should accept FileRenameInfoEx.
The documentation is extremely sparse but it would appear that
SetFileInformationByHandle should handle FileRenameInfoEx in basically
the same manner as FileRenameInfo, the only difference between the two
being how they store the flag for replacing an existing file.

No other flags are being tested as even on MSDN they're only properly
explained for NtSetInformationFile, which is wrapped by
SetFileInformationByHandle in Wine anyway.