d3dx9/tests: Fix volume test boxes.
This test seems to be testing different filters when creating volumes, and
nothing indicates it's intentionally testing volume creation from out-of-bound
srcboxes.

The pixels array is 16x4 bytes, pixel format D3DFMT_A8R8G8B8 is 4 bpp, so pixels
have 16 pixels in total. The boxes have 2 slices, with a slice pitch of 32
bytes, that's 8 pixel/slice. The original width*height is 6*6 which is thus
beyond these bounds. This commit changed it to 4*2 so the rowpitch of 16 bytes
is perserved.