mmdevapi/tests: Test extensible wave formats when rendering.
This adds a few more cases in which initializing in exclusive
mode can fail even when IsFormatSupported() returned positively.
To avoid making the broken() conditions overly complicated I
remove the side conditions, which are probably not very useful
and depend on the driver anyway.

Also, testing configurations with more than two channels somehow
alters the state of the device in a way that is not restored
even after releasing the IMMDevice object and creating it again.
Specifically, GetChannelVolume() will return more then two
channels, even if the audio client is initialized with just two
channels. To avoid those failures the format tests are moved
to the end.