win32u: Clear DC opengl drawable when releasing cached dce.
We won't be able to do it later on otherwise as get_dc_ptr will fail
with a disabled DC, and release_dce will then not clear the drawable.

This will delay the drawable release until the dce is later reused, for
instance for a different window, which may cause a complicated deadlock.