winegstreamer: Only add the capsfilter for avdec_h264.
h264parse currently has a bug where it will send an avc caps without
codec_data when it has received an SPS but not PPS.

This bug currently causes avdec_h264 to fail when a drain request
is made during caps neg.

As a workaround, a capsfilter was added to force a stream-format of
byte-stream for all h.264 decoders. However, this introduced a new
issue with the vtdec element, as it only supports the avc stream-format.

Therefore, this patch looks to address both problems by only introducing
the capsfilter when the avdec_h264 element is used.

It also removes the stream-format and alignment requirements from
the parsed caps, applying them only to the capsfilter.