winebus: Override device instance enumerator string if bus type is known.
Some games use the bus device instance to query properties of a HID
device. Tekken 8 does this to get a string for the device to display in
the UI. It queries all present devices for their container IDs, and when
it finds one with the same container ID as the HID device, it checks the
device instance string for a "BTHENUM" or "USB" enumerator.

Signed-off-by: Connor McAdams <cmcadams@codeweavers.com>