winebus: Unspecified condition blocks are full strength.
Under Windows and the PID specification it is possible to specify
less than the full set of condition blocks. Under SDL and the
Linux FF input API all condition blocks must always be specified.

This sets unspecified axes in the SDL and input backends to play
at full strength to agree with Windows and hidraw.