joy.cpl: Play condition effects on indicated axis too.
Testing with a MS Sidewinde 2 under Windows reveals that playing
a codition effect on just one axis can only be done by explicitly
setting the other axes to a zero strength as unspecified axes
run at full strength and not no strength as might be expected.

The choosen axis must also be the Y axis in order for the result
of the rotation to agree with what happens with non-condition
effects. This is because the the -Y axis is the reference axis
for the direction (i.e., no rotation plays back on the Y axis).