ntdll: Initialize return value in fork_and_exec.
Otherwise, if the exec succeeds, the later read() will not actually
read anything (EOF since the other end of the pipe is closed), and
we'll returned the uninitialized value of `status`.

This was preventing NtCreateUserProcess for a Unix process from
returning successfully, depending on the state of the stack, even
if the fork/exec suceeded.