msxml3/element: Implement removeAttributeNode function.
Implement the function `IXMLDOMElement_removeAttributeNode`.
Check if the given attribute has the element as parent. If so remove it
just like `domelem_remove_qualified_item` does.

Use the fact, that attribute names are unique in elements.

Add unittest to check if `removeAttributeNode` with output variable set
to `NULL` still removes the attribute from the element.