server: Do not clear queue masks in msg_queue_satisfied.
The masks are checked in win32u, and updated if necessary, right before
waiting for messages. They tell wineserver which of the bits we want the
queue to be signaled for in the following wait.

There should be no need to clear the mask after waiting, and doing it
will always force clients to update them again on the next wait, causing
an additional request.