comctl32: Remove syslink from comctl32 v5.
test_comctl32_classes() in comctl32 tests shows that syslink is not available in comctl32 v5. The
syslink test in the test_comctl32_classes() in user32 tests is removed in this commit because it's
not being tested properly. The broken() call at class.c#L1329 is hiding the failure on Windows.