dsound: Do not query for mixing format support.
Despite its name, IsFormatSupported() doesn't really predict what
Initialize() will return, especially when AUTOCONVERTPCM is used.
So there's no point querying it, let's just call Initialize() and
that will tell whether the format is accepted or not.

It is also useless to change to the recommended format: both if
we're going to use a primary buffer or not, the format that is
computer in DSOUND_WaveFormat() is what everything else expects,
and if it cannot be satisfied there's not point continuing.