winex11: Prevent larger than monitor window resizes.
This prevents WMs like KWin or Mutter from automatically maximizing
windows which frames are larger than their monitor area.

The maximization otherwise gets fed back into the win32 state, and often
confuses applications which are operating a fullscreen window change by
resizing a window then later changing its decoration.

It isn't possible to tell between a maximization initiated by the user
and one that is initiated by the WM responding to a window resize from
us. This heuristic should hopefully be good enough, as it is unlikely
that the user will ever be able to resize a window in a such a way that
its *frame* covers a monitor entirely. Note that if the user resizes a
window such that its *visible* area covers a monitor entirely, we are
going to make it fullscreen already anyway.