ntdll: Handle . and .. segments in relative symlinks.
Note that we don't want to use collapse_path() as-is, because everything else it
does (i.e. forward slashes, duplicate slashes, trailing dots) should explicitly
not apply to symlinks. We also should fail when rewinding past the drive root,
rather than simply ignoring those .. segments.

The FIXME introduced here related to unwinding past the RootDirectory handle is
not particularly problematic to address (we just need to query the server for
its whole path). I've left it as a FIXME simply because it takes extra work to
implement and is unlikely to come up; RootDirectory not used anywhere in
kernel32 or kernelbase file code.