comctl32: Use a magic value to check if an image list is valid.
Use a magic value to check if an image list is valid instead of using vtable addresses. So that an
image list from comctl32 v5 can be used with comctl32 v6 image list functions and vice versa.

Fix missing icons in the listview window of the Word 2010 file open dialog.