comctl32/ipaddress: Use the system window text color to draw dots.
Manual tests on XP show that ipaddress controls don't use the text colors in the theme file to draw
dots and the system window text color should be used instead. In fact, there is no text color property
for the ETS_NORMAL state in the theme file. This fixes the wrong color for the dots in ipaddress
controls.