server: Set NULL foreground input when switching to desktop window.
When called from the application.

Changing foreground window to the desktop seems like a bad idea, but it
also seem to allow later foreground changes back to application windows,
as many D3D tests suggest.

Currently we translate these calls into requests to activate the root
window, which will usually have no effect, so allowing further changes
should be safe.

When the internal flag is set, we actually change the foreground to the
desktop window, which should later forbid any further changes made from
the application side, as this happens when a non-Wine or foreign window
has been focused.