winex11: Flush after presenting.
The following cause an image to be immediately visible on Windows:

* Direct3D Present()

* wglSwapBuffers() [after drawing to the back buffer]

* glFlush() [after drawing to the front buffer. Note that glFlush() is both
  necessary and sufficient.]

* vkQueuePresent()

This is all of the paths that ultimately call client_surface_present().

In order to ensure that pixels are immediately visible on Wine as well, we need
to flush immediately after presenting.

Based on a patch by Henri Verbeet.