ntdll/tests: Update the reference code for a changed test.
This test case was updated in 41d7baa0139252f66042313f366b199e710f7b7b.
The behaviour of this testcase has changed; the older behaviour
could still be reproduced on Windows 11 version 10.0.22000.2176,
but on 10.0.26100.6899 we see the newer behaviour. (As the test
was updated in Feb 2024, it probably changed in some earlier build
than that.)

Update the function definition to match the actual observed
behaviour.