ntdll: Rewrite arm64 packed unwind info handling.
Instead of trying to handcraft all the cases for dealing with
partial prologues, just synthesize the full unwind info for the
function and execute that instead. The packed unwind info format
always maps to a combination of regular unwind opcodes.

This is how the arm packed unwind info implementation works
already; this makes it easier to reason about compared with the
handwritten code for dealing with partial prologues.

The previous implementation is certainly faster and more efficient,
but much harder to reason about, and to modify.