wined3d: Remove now unnecessary pixel format restoration.
wglSetPixelFormatWINE now sets the internal pixel format privately on
each DC, and it is unnecessary to restore the window pixel format.

That call is even undesirable as it allocates and destroys a private
OpenGL drawable when the DC is acquired, its internal pixel format is
set, and the DC is then immediately released.