twaindsm: Put DSM code into twaindsm.dll and add twain_32.dll wrapper.
Renamed dlls/twain_32 to dlls/twaindsm thus creating twaindsm.dll as
platform independent name for the Twain Data Source Manager
as it is used on 64-Bit windows.

Create new dlls/twain_32 as a wrapper for 32-Bit Windows only that
forwards all calls to the main implementation in twaindsm.dll.

Send the DG_CONTROL/DAT_ENTRYPOINT/MSG_SET before DG_CONTROL/DAT_IDENTITY/MSG_OPENDS
so Twain 2.x aware data sources like sane.ds don't need to know the
filename of the Twain DSM to find the DSM entry point.

On 32-Bit Systems LoadLibrary for the twain_32.dll wrapper when
opening the DS and it is not 2.x compliant. This makes sure
it will find the DSM_Entry of twain_32.dll if it uses
GetModuleHandle("twain_32.dll") to find the DLL.