wined3d: Update the context DC from its swapchain if it changed.
The context keeps a weak pointer to the swapchain, and the swapchain
might be destroyed while are context still have a pointer to them.

It's usually not causing issues because contexts will be refreshed with
the new active swapchain before a present is executed. However, the new
swapchain might be for the same window but might have a different DC.

As we now borrow the DCs from the swapchains, we also need to check if
the context DC needs to be updated.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=58709