server: Remove now unnecessary process-wide set_foreground flag.
We allow windows to be each activated at least once after their creation
and this is now mostly unnecessary. It also effectively allows processes
to activate their first window twice, once because the window was never
active before, and a second time because we only set the process-wide
flag after checking the window flag, which is not the intended effect.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=58167